Don't modify DHCP response when checking authentication
We can call md5_hash() as many times as we like, so pass in the
response as two parts with a zero hash in between rather than actually
modifying the response.

So far as I can see, the server hash is only ever used once, but
the current behaviour is surprising and violates const-correctness.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>